Fixes for VSrc-Implementation

+++ software/firmware/pru0-shepherd-fw/include/virtual_storage.h
@@ -9,6 +9,6 @@ void     storage_initialize();
 uint32_t get_SoC_1_n30(void); // for testing
 uint32_t get_V_OC_uV(void);   // for testing
 
-uint32_t storage_update(const uint64_t, const bool_ft);
+uint32_t storage_update(const uint32_t, const bool_ft);
 
 #endif //VIRTUAL_STORAGE_H
